Refactor components to include underconstruction text and link to products
