--- pages/Underconstruction.tsx
@@ -1,24 +1,0 @@
-import Head from 'next/head';
-import Link from 'next/link';
-import { underConstruction } from '../components/elements';
-import Layout from '../components/Layout';
-import { disableGrayLayer } from '../hooks';
-
-export default function UnderConstruction() {
-  // disableGrayLayer(props.showGrayLayer, props.setShowGrayLayer);
-  return (
-    <Layout>
-      <Head>
-        <title>under construction</title>
-        <meta name="description" content="site is under construction" />
-      </Head>
-
-      <section css={underConstruction}>
-        <h1>Under Construction</h1>
-        <Link href="/products" passHref>
-          <button>Continue Shopping</button>
-        </Link>
-      </section>
-    </Layout>
-  );
-}
